+++ tests/testthat/test-is-alluvial.r
@@ -1,16 +1,13 @@
 library(ggalluvial)
-context("is_alluvial")
+context("alluvial-data")
 
 titanic_alluvia <- as.data.frame(Titanic)
 
-test_that("is_alluvial recognizes alluvia-format Titanic data", {
-  expect_message(is_alluvial(titanic_alluvia),
-                 "[Mm]issing")
-  expect_true(is_alluvial(titanic_alluvia, axes = 1:4))
-  expect_true(is_alluvial(titanic_alluvia, axes = c("Class", "Sex")))
-  expect_warning(is_alluvial(titanic_alluvia,
-                             key = "Class", value = "Freq", id = "Age"),
-                 "[Dd]uplicate")
+# updated `is_alluvia_form()` tests
+test_that("`is_alluvia_form()` recognizes alluvia-format Titanic data", {
+  expect_message(is_alluvia_form(titanic_alluvia), "[Mm]issing")
+  expect_true(is_alluvia_form(titanic_alluvia, axes = 1:4))
+  expect_true(is_alluvia_form(titanic_alluvia, axes = c("Class", "Sex")))
 })
 
 titanic_lodes <- suppressWarnings(tidyr::gather(
@@ -21,14 +18,12 @@ titanic_lodes$Value <- factor(titanic_lodes$Value,
                               levels = do.call(c, lapply(titanic_alluvia[, 1:4],
                                                          levels)))
 
-test_that("is_alluvial recognizes lodes-format Titanic data", {
-  expect_message(is_alluvial(titanic_lodes),
-                 "[Mm]issing")
-  expect_true(is_alluvial(titanic_lodes,
-                          key = "Variable", value = "Value", id = "Index"))
-  expect_true(is_alluvial(titanic_lodes,
-                          key = "Variable", value = "Value", id = "Index",
-                          weight = "Freq"))
-  expect_message(is_alluvial(titanic_lodes, axes = 3:4),
-                 "[Mm]issing")
+# updated `is_alluvia_form()` tests
+test_that("`is_lodes_form()` recognizes lodes-format Titanic data", {
+  expect_error(is_lodes_form(titanic_lodes))
+  expect_true(is_lodes_form(titanic_lodes,
+                            key = Variable, value = Value, id = Index))
+  expect_true(is_lodes_form(titanic_lodes,
+                            key = "Variable", value = "Value", id = "Index",
+                            weight = "Freq"))
 })